게시글 수정 일부 기능 개선

+++ src/main/java/com/one/whospet/service/board/impl/BoardServiceImpl.java
@@ -313,9 +313,11 @@ public class BoardServiceImpl implements BoardService {
 		
 		//유저가 삭제선택한 이미지 정보가 있는지 체크
 		int count = 1;
+		boolean flag = true;
 		while(true) {
 			String fileInfo = fileRequest.getParameter("fileInfo" + count);
 			if(fileInfo != null && !fileInfo.equals("")) {
+				flag = false; //사용자 선택삭제 기능 사용
 				boardDao.deleteCheckFile(fileInfo); // db에서 선택된파일지움			
 				String path = context.getRealPath("upload");	
 				//현재 게시판에 존재하는 파일객체를 만듬
@@ -330,35 +332,36 @@ public class BoardServiceImpl implements BoardService {
 			}
 		}
 		
-		
-		
-		
+			
 		//첨부파일이 null이거나 없으면 게시글 파일이 없으면 작성 끝
 		if(fileList == null || fileList.isEmpty() || fileList.get(0).getSize() == 0) {
 			logger.info("파일 없음");
 			return;
 		}else {
-			logger.info("파일 있음");
-			//파일이 있는 경우
-			//삭제할 파일 정보를 가져온다.
-			List<BoardImg> list = boardDao.deleteFileInfo(bNo);
-			
-			//저장된 파일정보가 null이 아니고 리스트가 비어있지 않으면 실행
-			if(list != null && !list.isEmpty()) {
-				for(int i = 0; i < list.size(); i++) {
-					//파일 경로 지정
-					String path = context.getRealPath("upload");
-					
-					//현재 게시판에 존재하는 파일객체를 만듬
-					File file = new File(path + "\\" + list.get(i).getBiStoredFilename());
-					
-					if(file.exists()) { // 파일이 존재하면
-						file.delete(); // 파일 삭제	
-					}
-				}
+			if(flag) {//사용자가 선택삭제 기능을 사용 안함
+				logger.info("파일 있음");
+				//파일이 있는 경우
+				//삭제할 파일 정보를 가져온다.
+				List<BoardImg> list = boardDao.deleteFileInfo(bNo);
 				
-				//저장된 파일을 모두 제거한 후 DB 에서도 지운다.
-				boardDao.deleteBoardFile(bNo);
+				//저장된 파일정보가 null이 아니고 리스트가 비어있지 않으면 실행
+				if(list != null && !list.isEmpty()) {
+					for(int i = 0; i < list.size(); i++) {
+						//파일 경로 지정
+						String path = context.getRealPath("upload");
+						
+						//현재 게시판에 존재하는 파일객체를 만듬
+						File file = new File(path + "\\" + list.get(i).getBiStoredFilename());
+						
+						if(file.exists()) { // 파일이 존재하면
+							file.delete(); // 파일 삭제	
+						}
+					}
+					
+					//저장된 파일을 모두 제거한 후 DB 에서도 지운다.
+					boardDao.deleteBoardFile(bNo);
+				}
+			
 			}	
 			
 			//-------------- 새로운 파일을 등록 하는 코드
